Move generic big endian functions to common
